Removed the use of module header files
Only the .cc files need the module declaration.
