+++ GeneratorInterface/Core/plugins/GenFilterEfficiencyAnalyzer.cc
@@ -1,5 +1,46 @@
-#include "GeneratorInterface/Core/interface/GenFilterEfficiencyAnalyzer.h"
+// F. Cossutti
+// $Revision://
+
+// analyzer of a summary information product on filter efficiency for a user specified path
+// meant for the generator filter efficiency calculation
+
+// system include files
+#include <memory>
 #include <iostream>
+
+// user include files
+
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "FWCore/Framework/interface/EDAnalyzer.h"
+
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/Run.h"
+#include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/Framework/interface/LuminosityBlock.h"
+#include "FWCore/Utilities/interface/EDGetToken.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+
+#include "SimDataFormats/GeneratorProducts/interface/GenFilterInfo.h"
+//
+// class declaration
+//
+
+class GenFilterEfficiencyAnalyzer : public edm::EDAnalyzer {
+public:
+  explicit GenFilterEfficiencyAnalyzer(const edm::ParameterSet&);
+  ~GenFilterEfficiencyAnalyzer() override;
+
+private:
+  void analyze(const edm::Event&, const edm::EventSetup&) override;
+  void endLuminosityBlock(edm::LuminosityBlock const&, edm::EventSetup const&) override;
+  void endJob() override;
+
+  edm::EDGetTokenT<GenFilterInfo> genFilterInfoToken_;
+  GenFilterInfo totalGenFilterInfo_;
+
+  // ----------member data ---------------------------
+};
 
 GenFilterEfficiencyAnalyzer::GenFilterEfficiencyAnalyzer(const edm::ParameterSet& pset)
     : genFilterInfoToken_(consumes<GenFilterInfo, edm::InLumi>(pset.getParameter<edm::InputTag>("genFilterInfoTag"))),
@@ -30,3 +71,5 @@ void GenFilterEfficiencyAnalyzer::endJob() {
   std::cout << "Filter efficiency = " << totalGenFilterInfo_.filterEfficiency(-1) << " +- "
             << totalGenFilterInfo_.filterEfficiencyError(-1) << std::endl;
 }
+
+DEFINE_FWK_MODULE(GenFilterEfficiencyAnalyzer);
