+++ GeneratorInterface/Core/plugins/GeneratorSmearedProducer.cc
@@ -1,4 +1,3 @@
-#include "GeneratorInterface/Core/interface/GeneratorSmearedProducer.h"
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/Utilities/interface/InputTag.h"
 #include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
@@ -6,7 +5,31 @@
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 
 #include "SimDataFormats/GeneratorProducts/interface/HepMCProduct.h"
+#include "FWCore/Framework/interface/global/EDProducer.h"
+#include "FWCore/Utilities/interface/EDGetToken.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+
 #include <memory>
+
+namespace edm {
+  class ParameterSet;
+  class ConfigurationDescriptions;
+  class Event;
+  class EventSetup;
+  class HepMCProduct;
+}  // namespace edm
+
+class GeneratorSmearedProducer : public edm::global::EDProducer<> {
+public:
+  explicit GeneratorSmearedProducer(edm::ParameterSet const& p);
+
+  void produce(edm::StreamID, edm::Event& e, edm::EventSetup const& c) const override;
+  static void fillDescriptions(edm::ConfigurationDescriptions& descriptions);
+
+private:
+  const edm::EDGetTokenT<edm::HepMCProduct> newToken_;
+  const edm::EDGetTokenT<edm::HepMCProduct> oldToken_;
+};
 
 GeneratorSmearedProducer::GeneratorSmearedProducer(edm::ParameterSet const& ps)
     : newToken_(consumes<edm::HepMCProduct>(ps.getUntrackedParameter<edm::InputTag>("currentTag"))),
@@ -34,3 +57,5 @@ void GeneratorSmearedProducer::fillDescriptions(edm::ConfigurationDescriptions& 
   desc.addUntracked<edm::InputTag>("previousTag", edm::InputTag("generator"));
   descriptions.add("generatorSmeared", desc);
 }
+
+DEFINE_FWK_MODULE(GeneratorSmearedProducer);
