--- GeneratorInterface/Core/plugins/module.cc
@@ -1,10 +1,0 @@
-#include "FWCore/Framework/interface/MakerMacros.h"
-
-#include "GeneratorInterface/Core/interface/GenFilterEfficiencyAnalyzer.h"
-DEFINE_FWK_MODULE(GenFilterEfficiencyAnalyzer);
-
-#include "GeneratorInterface/Core/interface/GenXSecAnalyzer.h"
-DEFINE_FWK_MODULE(GenXSecAnalyzer);
-
-#include "GeneratorInterface/Core/interface/GeneratorSmearedProducer.h"
-DEFINE_FWK_MODULE(GeneratorSmearedProducer);
